feat(input): add support for input icons and different input sizes
This commit adds support for icons in the input component. It also adds the ability to change the size of the input component. The input component now has two sizes: "small" and "base". The "variant" prop has been removed and replaced with the "size" prop. The "size" prop can be set to "small" or "base". The "icon" prop can be used to add an icon to the input component.