feat(divider): add Divider component with text and textPosition props
The Divider component is a new component that can be used to separate content
with a horizontal line. It has two optional props: text and textPosition. The
text prop is used to add a text to the divider, and the textPosition prop is
used to define the position of the text. The textPosition prop can be set to
"top-left" or "middle". If the textPosition is set to "top-left", the text will
be displayed above the divider. If the textPosition is set to "middle", the
text will be displayed in the middle of the divider. If the text prop is not
set, the Divider component will only display a horizontal line.

The Divider component has been added to the atoms folder, and it has a story
file, a test file, a type file, and an index file. The tailwind.config file
has also been